% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustertree.R
\name{clustertree}
\alias{clustertree}
\title{Compute the Empirical Cluster Tree}
\usage{
clustertree(x, d = ncol(x), k = "suggest", alpha = "suggest",
  estimator = c("RSL", "knn", "mutual knn"), warn = FALSE)
}
\arguments{
\item{x}{a matrix-coercible data set, or a euclidean 'dist' object.}

\item{k}{integer; smoothing parameter, controls the radius of the ball containing k points around each x_i.}

\item{alpha}{float; regularity parameter determing when to connect points. See below for details.}
}
\description{
More details coming soon...
}
\note{
The default 'k' parameter will differ between if a 'dist' object is given vs. the original data set, as with
the dist object the original dimensionality of the data is unknown (to which the default setting of k depends on).
}
\references{
See KC and SD.
}
