% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distortion.R
\name{distortion}
\alias{distortion}
\title{Merge distortion}
\usage{
distortion(hc1, hc2, correspondence = "identity")
}
\arguments{
\item{hc1}{hclust object; the first hierarchy}

\item{hc2}{hclust object; the second hierarchy}

\item{correspondence}{the correspondence function to use.}
}
\description{
Implements the merge distortion metric described in [1]. Currently only supports the identity correspondence function.
}
\references{
1. Eldridge, Justin, Mikhail Belkin, and Yusu Wang. "Beyond hartigan consistency: Merge distortion metric for hierarchical clustering." Conference on Learning Theory. 2015.
}
