% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{K-nearest neighbor search}
\usage{
knn(x, k, r_x = NULL, bucketSize = 15L, splitRule = "suggest",
  approx = 0, ...)
}
\arguments{
\item{x}{a matrix-coercible data object.}

\item{k}{the number of neighbors to find.}

\item{r_x}{the reference data set, if any. See details.}

\item{bucketSize}{maximum size of the kd-tree leaves.}

\item{splitRule}{rule to split the kd-tree. One of "STD", "MIDPT", "FAIR", "SL_MIDPT", "SL_FAIR" or "SUGGEST" (SL stands for sliding).}

\item{approx}{whether to allow an arbitrary small amount of error in the result for better speed.}

\item{...}{unused.}
}
\description{
Performs a k-nearest neighbor search.

"SUGGEST" uses the ANN authors preferred 'best' rule. If no reference data set (\code{rx}) is specified,
then the k-nearest neighbors in \code{x} are found. If a reference data set is specified, a kd tree is built using the
reference set, and then the k-nearest reference neighbors to each query pt are returned.
}
\references{
1. Mount, David M., and Sunil Arya. "ANN: library for approximate nearest neighbour searching." (1998).
}
