% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustertree.R
\name{plot.clustertree}
\alias{plot.clustertree}
\title{Plot a given cluster tree}
\usage{
\method{plot}{clustertree}(C_n, type = c("both", "dendrogram", "span tree"))
}
\arguments{
\item{x}{a 'clustertree' object.}
}
\description{
More details coming soon...
}
\references{
See KC and SD.
}
