% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustertree.R
\name{print.clustertree}
\alias{print.clustertree}
\title{print.clustertree}
\usage{
\method{print}{clustertree}(x, ...)
}
\arguments{
\item{x}{a 'clustertree' object.}

\item{...}{unused.}
}
\description{
Prints various parameters used to compute the cluster tree.
}
