% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Prune an empirical cluster tree}
\usage{
prune(C_n, delta, eps = 1/sqrt(C_n$k))
}
\arguments{
\item{delta}{The (1-δ) probability}

\item{eps}{A tuning parameter (> 0) representing how aggressively to prune.}
}
\description{
Implementation of Chaudhuri et. al's consistent pruning procedure for the cluster tree. By "pruning" the tree,
spurious clusters are removed while salient clusters are recovered.
}
\references{
See KC and SD (2014).
}
