% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runt_prune.R
\name{runt_prune}
\alias{runt_prune}
\title{Runt Pruning}
\usage{
runt_prune(C_n, runt_size)
}
\arguments{
\item{C_n}{Either a 'clustertree' or a generic 'hclust' object.}

\item{runt_size}{Positive integer. The number of original observations a branch needs to contain to be considered "real" or "significant". See details.}
}
\value{
If a 'clustertree' object was given, the underlying hierarchy (or hierarchies) are replaced by pruned hierarchies. If a regular
'hclust' is given, a simplified (runt-pruned) version of the hierarchy is returned.
}
\description{
Implementation of Stuetzle's 'Runt pruning' for cluster tree and generic hierarchical cluster representations.

Runt pruning simplifies a given cluster tree into a new, pruned cluster tree. Algorithmically, it
begins by divisively traversing a given cluster tree, and at each split, looking at
whether a branch contains a sufficient number of observations (determined by the \code{runt_size}). If it does, the
branch (and its associated connected component) is considered 'significant', and is retained in the simplified tree.
If it does not, the entire branch and its descendents are grouped together, and are collectively recorded as either:
1) A leaf, if the other branch also has less than then \code{runt_size} number of observations.
2) As part of the splitting branch, if the other branch has at least \code{runt_size} number of observations.
}
\references{
1. Stuetzle, Werner. "Estimating the cluster tree of a density by analyzing the minimal spanning tree of a sample." Journal of classification 20.1 (2003): 025-047.
}
\seealso{
cut.simplified_hclust
}
