% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustertree.R
\name{clustertree}
\alias{clustertree}
\title{Compute the Empirical Cluster Tree}
\usage{
clustertree(x, k = "suggest", alpha = "suggest", estimator = c("RSL",
  "knn", "mutual knn"), warn = FALSE)
}
\arguments{
\item{x}{a matrix-coercible data set, or a euclidean 'dist' object.}

\item{k}{integer; smoothing parameter, controls the radius of the ball containing k points around each x_i.}

\item{alpha}{float; regularity parameter that prevents excessive chaining. Can be safely kept on the default. See below for details.}
}
\description{
Implements various estimators for estimating the empirical cluster tree.
}
\note{
The default 'k' parameter will differ between if a 'dist' object is given vs. the original data set, as with
the dist object the original dimensionality of the data is unknown (to which the default setting of k depends on).
Chaudhuri et. al recommend keeping alpha as low as possible, subject to being no lower than sqrt(2).
}
\references{
See Chaudhuri, Kamalika, and Sanjoy Dasgupta. "Rates of convergence for the cluster tree." Advances in Neural Information Processing Systems. 2010.
}
