% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustertree.R
\name{clustertree}
\alias{clustertree}
\title{Compute the Empirical Cluster Tree}
\usage{
clustertree(x, k = "suggest", alpha = "suggest", estimator = c("RSL",
  "knn", "mutual knn"), warn = FALSE)
}
\arguments{
\item{x}{a matrix-coercible data set, or a euclidean 'dist' object.}

\item{k}{integer; smoothing parameter controlling how 'dense' an object should be before being considered for
linkage into the cluster tree hierarchy. See below for details.}

\item{alpha}{float; regularity parameter used to prevent 'excessive chaining'. Can be safely kept on the default. See below for details.}

\item{estimator}{Which estimator to choose from. Default is robust single linkage, see below for details.}

\item{warn}{boolean; Whether to warn the user regarding the usage of the parameter settings. Defaults to false. See below for details.}
}
\description{
Implements various algorithms for estimating the cluster tree.
}
\note{
The default 'k' parameter will differ between if a 'dist' object is given vs. the original data set, as with
the dist object the original dimensionality of the data is unknown (to which the default setting of k depends on).
Chaudhuri et. al [1] recommend keeping alpha as low as possible, but recommended no lower than sqrt(2). If not estimator
is specified, Robust Single Linkage (RSL) is used.

The other two estimators are from Algorithm 2 in [2]. Note that in the case where the KNN or Mutual KNN graph estimators are used,
the resulting 'tree' may actually be a disconnected forest. To handle this each connected component is made into its
own hierarchy ('hclust' object), discarding singletons (if any). Because \code{hclust} objects require a representation
where leaf indices start with 1, the indices of the original data set are stored in the \code{idx} member of each
hierarchy.
}
\references{
1. Chaudhuri, Kamalika, and Sanjoy Dasgupta. "Rates of convergence for the cluster tree." Advances in Neural Information Processing Systems. 2010.
2. Chaudhuri, Kamalika, et al. "Consistent procedures for cluster tree estimation and pruning." IEEE Transactions on Information Theory 60.12 (2014): 7900-7912.
}
\seealso{
hclust
}
