% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cons_prune.R
\name{cons_prune}
\alias{cons_prune}
\title{Consistently prune the empirical cluster tree}
\usage{
cons_prune(C_n, delta, eps = 1/sqrt(C_n$k))
}
\arguments{
\item{C_n}{A clustertree object.}

\item{delta}{The (1-\deqn{\delta}) probability threshold associated with the pruning constant. See below for details.}

\item{eps}{A tuning parameter (> 0) representing how aggressively to prune.}
}
\description{
DO NOT USE. Current unfinished.

Implementation of Chaudhuri et. al's consistent pruning procedure for the cluster tree. By "pruning" the tree,
spurious clusters are removed while salient clusters are recovered.
}
\references{
1. Chaudhuri, Kamalika, et al. "Consistent procedures for cluster tree estimation and pruning." IEEE Transactions on Information Theory 60.12 (2014): 7900-7912.
}
