% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustertree.R
\name{plot.clustertree}
\alias{plot.clustertree}
\title{Plot a given cluster tree}
\usage{
\method{plot}{clustertree}(C_n, which = 1, h = NULL, f = "scatter")
}
\arguments{
\item{which}{if there are several hierarchies from the original data set, which one to plot?}

\item{x}{a 'clustertree' object.}
}
\description{
More details coming soon...
}
\references{
See KC and SD.
}
