% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.clustertree}
\alias{plot.clustertree}
\title{Plot a given cluster tree}
\usage{
plot.clustertree(x)
}
\arguments{
\item{x}{a 'clustertree' object.}
}
\description{
More details coming soon...
}
\references{
See KC and SD.
}
