% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanplot.R
\name{spanplot}
\alias{spanplot}
\title{Plot a spanning tree}
\usage{
spanplot(X_n, C_n, h = NULL)
}
\arguments{
\item{X_n}{a matrix to plot as a scatter plot.}

\item{C_n}{a 'clustertree' object}

\item{h}{an optional height to cut the tree at. Colors the corresponding connected components.}
}
\description{
Allows for visualizing a clustertree object as a spanning tree.
}
